kernel: Remove interface thread_wait
Now sys_timer is always allocated on stack and register for thread killing
