+++ src/kernel/thread/thread_sched_wait.c
@@ -42,10 +42,35 @@ static void sched_wait_timeout_handler(struct sys_timer *timer, void *data) {
 	sched_wakeup(s);
 }
 
+
+void thread_wait_init(struct thread_wait *tw) {
+	dlist_init(&tw->thread_waitq_list);
+}
+
+static inline void thread_wait_add(struct thread_wait *tw, struct sys_timer *tmr) {
+	dlist_add_next(&tw->thread_waitq_list, dlist_init(&tmr->st_wait_link));
+}
+
+static inline void thread_wait_del(struct sys_timer *tmr) {
+	dlist_del(&tmr->st_wait_link);
+}
+
+void thread_wait_deinit(struct thread_wait *tw) {
+	struct sys_timer *tmr;
+
+	dlist_foreach_entry_safe(tmr, &tw->thread_waitq_list, st_wait_link) {
+		timer_close(tmr);
+	}
+	dlist_init(&tw->thread_waitq_list);
+}
+
 int sched_wait_timeout(clock_t timeout, clock_t *remain) {
-	SCHED_WAIT_TIMER_DEF(tmr);
+	struct sys_timer tmr;
 	clock_t remain_v, cur_time;
 	int res, diff;
+	struct thread *thr;
+
+	thr = thread_self();
 
 	if (timeout == SCHED_TIMEOUT_INFINITE) {
 		remain_v = SCHED_TIMEOUT_INFINITE;
@@ -54,15 +79,18 @@ int sched_wait_timeout(clock_t timeout, clock_t *remain) {
 	}
 
 	cur_time = clock();
-	SCHED_WAIT_TIMER_ADD(tmr);
-	if ((res = SCHED_WAIT_TIMER_INIT(tmr, TIMER_ONESHOT, jiffies2ms(timeout),
+	thread_wait_add(&thr->thread_wait_list, &tmr);
+	if ((res = timer_init_start_msec(&tmr, TIMER_ONESHOT, jiffies2ms(timeout),
 			sched_wait_timeout_handler, schedee_get_current()))) {
+		thread_wait_del(&tmr);
 		return res;
 	}
 
 	schedule();
 	diff = clock() - cur_time;
-	SCHED_WAIT_TIMER_CLOSE(tmr);
+
+	timer_close(&tmr);
+	thread_wait_del(&tmr);
 
 	if (diff < timeout) {
 		remain_v = timeout - diff;
