--- src/kernel/thread/thread_wait/thread_wait_stub.c
@@ -1,21 +1,0 @@
-/**
- * @file
- * @brief
- *
- * @author  Anton Kozlov
- * @date    29.05.2014
- */
-
-#include <kernel/thread/thread_wait.h>
-
-void thread_wait_init(struct thread_wait *tw) {
-}
-
-void thread_wait_add(struct thread_wait *tw, struct sys_timer *tmr) {
-}
-
-void thread_wait_del(struct thread_wait *tw, struct sys_timer *tmr) {
-}
-
-void thread_wait_deinit(struct thread_wait *tw) {
-}
