+++ templates/arm/qemu/mods.config
@@ -209,6 +209,4 @@ configuration conf {
 	@Runlevel(1) include embox.test.math.math_test
 	@Runlevel(2) include embox.test.math.fpu_context_consistency_test
 	@Runlevel(2) include embox.test.posix.fcntl.open_test
-
-	include embox.kernel.thread.thread_wait_on_stack
 }
