+++ templates/arm/stm32_vl_light/mods.config
@@ -41,7 +41,6 @@ configuration conf {
 	include embox.kernel.thread.thread_cancel_disable
 	include embox.kernel.thread.stack_none
 	include embox.kernel.thread.signal_stub
-	include embox.kernel.thread.thread_wait_stub
 	@Runlevel(2) include embox.kernel.sched.boot_light
 	@Runlevel(2) include embox.kernel.sched.timing.none
