+++ templates/x86/qemu/mods.config
@@ -193,6 +193,4 @@ configuration conf {
 	include third_party.dash
 	include embox.compat.posix.proc.fork_copy_everything
 	include embox.arch.x86.fork
-
-	include embox.kernel.thread.thread_wait_on_stack
 }
